refactor: Emphasise command module is for external cli calls
