+++ src/check/rustup_toolchain_check.rs
@@ -1,9 +1,9 @@
 use crate::check::Check;
-use crate::command::cargo_command::CargoCommand;
-use crate::command::rustup_command::RustupCommand;
 use crate::context::EnvironmentContext;
 use crate::download::{DownloadToolchain, ToolchainDownloader};
 use crate::error::{IoError, IoErrorSource};
+use crate::external_command::cargo_command::CargoCommand;
+use crate::external_command::rustup_command::RustupCommand;
 use crate::lockfile::LockfileHandler;
 use crate::reporter::event::{CheckMethod, CheckResult, CheckToolchain, Method};
 use crate::toolchain::ToolchainSpec;
